refactor cargo-shuttle to bin+lib for testing
